requirements/test.txt
# Packages required for testing
-r base.txt

coverage==4.5.1
django-dynamic-fixture==2.0.0
django-nose==1.4.5
edx-lint==0.5.5
mock==2.0.0
nose-ignore-docstring==0.2
pep8==1.7.1
